rename: WireGuardInterface 파일로 분리
